refactor: move elevator constants to its own file
