+++ src/main/java/org/team1540/robot2025/subsystems/elevator/Elevator.java
@@ -1,6 +1,6 @@
 package org.team1540.robot2025.subsystems.elevator;
 
-import static org.team1540.robot2025.Constants.Elevator.*;
+import static org.team1540.robot2025.subsystems.elevator.ElevatorConstants.*;
 
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj.DriverStation;
